requirements.txt
 Flask
 boto3
 python-dotenv
 requests
 cryptography
 qrcode[pil]
 Pillow
 gunicorn

